feat: add SDE updater, price history, and fix market scan region handling
- SDE update task: downloads CCP SDE ZIP, parses YAML, imports into DB
- Price history task: fetches ESI market history for top traded items
- Market scan: fix region_id propagation from scan context to DB
- Celery: register new tasks, add cleanup schedule

+++ backend/app/models/market.py
@@ -9,7 +9,7 @@ class MarketOrder(Base):
 
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     type_id: Mapped[int] = mapped_column(Integer, index=True)
-    station_id: Mapped[int] = mapped_column(Integer)
+    station_id: Mapped[int] = mapped_column(BigInteger)
     region_id: Mapped[int] = mapped_column(Integer, index=True)
     system_id: Mapped[int] = mapped_column(Integer)
     order_id: Mapped[int] = mapped_column(BigInteger, unique=True)
